requirements.txt
paho-mqtt==1.6.1
opencv-python==4.5.5.64
numpy==1.22.3
pygame==2.1.2
base64==3.7.0